Switch Android binary deps to be managed by CIPD.
Removing b/a/update_deps/update_third_party_deps.py script.
Replacing with a simple CIPD wrapper script.

Adding ensure files for all each target_os (this is a file used to
describe the package you want to download). Initially, this is just
Android.

Adding cipd.yaml files for all deps (this file was used to
create the CIPD packages), and can be modified/used in the future
to update these deps.

CIPD gives us the advantage of having a consistent method of
downloading/managing our binary dependencies. The goal is to move
as much as possible from adhoc scripts over to CIPD. CIPD does
various nice things, such as verifying the integrity of the fetched
binary and doing smart caching. What our adhoc scripts do currently
is very inconsistent.

Bug:

+++ build/android/gyp/copy_ex.py
@@ -30,7 +30,9 @@ def CopyFile(f, dest, deps):
     shutil.copytree(f, os.path.join(dest, os.path.basename(f)))
     deps.extend(_get_all_files(f))
   else:
-    shutil.copy(f, dest)
+    if not (os.path.exists(f) and os.path.exists(dest)
+            and os.path.samefile(f, dest)):
+      shutil.copy(f, dest)
     deps.append(f)
 
 def DoCopy(options, deps):
